+++ src/likelihood.c
@@ -47,7 +47,7 @@ int lrt(double nullmodel_logl, double ptp_logl, unsigned int df, double * pvalue
   return 1;
 }
 
-double aic(double logl, int k, int n)
+double aic(double logl, long k, long n)
 {
   if (k > 1) k++;
   return -2*logl + 2*k + (double)(2*k*(k + 1)) / (double)(n-k-1);
